made isMobile to a context and used useContext in all the files it was being used
