+++ components/NewsStartPage.jsx
@@ -3,23 +3,13 @@ import { Box } from "@mui/system";
 import Link from "next/link";
 import styles from "../styles/News.module.css";
 import { useRouter } from "next/router";
-import { useState, useEffect } from "react";
+import { useContext } from "react";
 import Image from "next/image";
+import { MobileContext } from "../pages/_app";
 
 const News = () => {
   const router = useRouter();
-  const [isMobile, setIsMobile] = useState(false);
-
-  useEffect(() => {
-    window.innerWidth < 900 ? setIsMobile(true) : setIsMobile(false);
-    const handleWindowResize = () => {
-      window.innerWidth < 900 ? setIsMobile(true) : setIsMobile(false);
-    };
-    window.addEventListener("resize", handleWindowResize);
-    return () => {
-      window.removeEventListener("resize", handleWindowResize);
-    };
-  }, []);
+  const isMobile = useContext(MobileContext)
 
   return (
     <>
